User endpoints, me endpoint for creating new user when registering. BlogController.cs fixes

+++ ReactBlog.Server/Controllers/BlogController.cs
@@ -12,21 +12,25 @@ namespace ReactBlog.Server.Controllers;
 [Route("blogs")]
 public class BlogController(IBlogService blogService, ILogger<BlogController> logger) : ControllerBase
 {
-    private string OwnerId => User.FindFirstValue("sub")!;
+    private string? OwnerId => User.FindFirstValue("sub");
 
     [HttpGet]
-    public async Task<IEnumerable<Blog>> GetBlogsAsync()
+    public async Task<ActionResult<IEnumerable<Blog>>> GetBlogsAsync()
     {
+        if (OwnerId is null) return Unauthorized();
+
         logger.LogInformation("Getting all blogs with OwnerId = {id}", OwnerId);
-        return await blogService.GetAllAsync(OwnerId);
+        return Ok(await blogService.GetAllAsync(OwnerId));
     }
 
     [HttpGet("{blogId:int}")]
     public async Task<ActionResult<Blog>> GetBlog(int blogId)
     {
+        if (OwnerId is null) return Unauthorized();
+
         var foundBlog = await blogService.GetBlogAsync(blogId, OwnerId);
         if (foundBlog != null) return Ok(foundBlog);
-        
+
         logger.LogInformation("Blog with id {blogId} not found.", blogId);
         return NotFound($"Blog with id {blogId} not found");
     }
@@ -34,6 +38,7 @@ public class BlogController(IBlogService blogService, ILogger<BlogController> lo
     [HttpPost]
     public async Task<ActionResult<Blog>> AddBlog([FromForm] NewBlogDto blogDto)
     {
+        if (OwnerId is null) return Unauthorized();
         if (blogDto == null || string.IsNullOrEmpty(blogDto.Name)) { return BadRequest(); }
 
         var newBlog = new Blog
@@ -47,17 +52,18 @@ public class BlogController(IBlogService blogService, ILogger<BlogController> lo
         };
 
         await blogService.AddBlog(newBlog);
-        
+
         return CreatedAtAction(nameof(GetBlog), new { blogId = newBlog.Id }, newBlog);
     }
 
     [HttpPut("{blogId:int}")]
     public async Task<ActionResult<Blog>> UpdateBlog(int blogId, [FromForm] NewBlogDto updatedBlog)
     {
+        if (OwnerId is null) return Unauthorized();
         if (updatedBlog == null) { return BadRequest(); }
-        
+
         var blog = await blogService.UpdateBlog(blogId, updatedBlog, OwnerId);
-        
+
         if (blog == null) return NotFound($"Blog {blogId} not found");
         return Ok(blog);
     }
@@ -65,9 +71,11 @@ public class BlogController(IBlogService blogService, ILogger<BlogController> lo
     [HttpDelete("{blogId:int}")]
     public async Task<IActionResult> RemoveBlog(int blogId)
     {
+        if (OwnerId is null) return Unauthorized();
+
         var deleted = await blogService.RemoveBlog(blogId, OwnerId);
         if (!deleted) { return NotFound($"Blog {blogId} not found"); }
-        
+
         return Ok("Blog deleted.");
     }
 }
